Update api version

+++ app/src/main/java/covid19/coronavirus/repository/CovidRepositoryImpl.kt
@@ -10,6 +10,8 @@ import covid19.coronavirus.model.CountryResponse
 import covid19.coronavirus.model.TotalResponse
 import covid19.coronavirus.util.NetworkUtil
 import covid19.coronavirus.util.formatDateFromMillis
+import retrofit2.HttpException
+import java.lang.Exception
 
 class CovidRepositoryImpl(
     private val context: Context,
@@ -22,14 +24,29 @@ class CovidRepositoryImpl(
         if (NetworkUtil.isOnline(context)) {
 
             val currentResponse = restApi.getCurrent()
-
-            database.countryDao().deleteAll()
-            database.countryDao().insertAll(currentResponse)
+            if (currentResponse.isSuccessful) {
+                currentResponse.body()?.let {
+                    database.countryDao().deleteAll()
+                    database.countryDao().insertAll(it)
+                }
+            } else {
+                if (database.countryDao().getCountry().isEmpty()) {
+                    throw HttpException(currentResponse)
+                }
+            }
 
             val totalResponse = restApi.getTotal()
-            totalResponse.id = totalResponse.javaClass.simpleName
-            sessionManager.lastDateUpdate = formatDateFromMillis(totalResponse.updated)
-            database.totalCasesDao().insert(totalResponse)
+            if (totalResponse.isSuccessful) {
+                totalResponse.body()?.let {
+                    it.id = totalResponse.javaClass.simpleName
+                    sessionManager.lastDateUpdate = formatDateFromMillis(it.updated)
+                    database.totalCasesDao().insert(it)
+                }
+            } else {
+                if (database.totalCasesDao().getTotalCases() == null){
+                    throw HttpException(totalResponse)
+                }
+            }
         } else {
             if (database.countryDao().getCountry().isEmpty()) {
                 throw NetworkException(context.getString(R.string.network_error_message))
@@ -50,7 +67,7 @@ class CovidRepositoryImpl(
     }
 
     override suspend fun getTotalCases(): TotalResponse {
-        return database.totalCasesDao().getTotalCases()
+        return database.totalCasesDao().getTotalCases()!!
     }
 
     override suspend fun getLastUpdate(): String {
